AP_HAL_ChibiOS: allow hwdef override of RX_BOUNCE_BUFSIZE
Guard the define and widen the idle-path length to uint16_t so bounce
buffers over 255 bytes are not truncated. A 1 KB bounce buffer cuts the
RX interrupt rate from ~20 kHz to ~1.2 kHz at 1.25 MB/s.

+++ libraries/AP_HAL_ChibiOS/UARTDriver.cpp
@@ -1147,7 +1147,7 @@ void UARTDriver::_rx_timer_tick(void)
         bool enabled = (rxdma->stream->CR & STM32_DMA_CR_EN);
 #endif
         if (!enabled) {
-            uint8_t len = RX_BOUNCE_BUFSIZE - dmaStreamGetTransactionSize(rxdma);
+            uint16_t len = RX_BOUNCE_BUFSIZE - dmaStreamGetTransactionSize(rxdma);
             if (len != 0) {
                 const uint32_t written = _readbuf.write(rx_bounce_buf[rx_bounce_idx], len);
                 _rx_stats_bytes += len;
